Moving the deploy class to be used with lambda
As we don't want to duplicate the code, we can refactor the deploy class
to be used by both the ecs-service and the lambda functions

JIRA: PLAT-885
