+++ cdflow_commands/cli.py
@@ -62,7 +62,6 @@ def _run(argv):
         global_config=global_config,
         root_session=root_session,
     )
-
     if args['release']:
         plugin.release()
     elif args['deploy']:
